requirements-local.txt
python-jobspy
fastapi
uvicorn[standard]
apscheduler
jinja2
aiofiles
libsql-client
